requirements.txt
fastapi
uvicorn[standard]
pymongo
PyJWT
python-dotenv
pydantic
